# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang application using Gin and GORM to interact with the Spotify API.",
        "title": "Spotify API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/track": {
            "post": {
                "description": "Create a new track record in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new track",
                "operationId": "create-track",
                "parameters": [
                    {
                        "description": "Track details to create",
                        "name": "trackDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TrackDetails"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.TrackDetails"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Track with ISRC code already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/track/{isrc}": {
            "get": {
                "description": "Get track details from the database or Spotify by ISRC code",
                "produces": [
                    "application/json"
                ],
                "summary": "Get track details by ISRC",
                "operationId": "get-track-by-isrc",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ISRC code of the track",
                        "name": "isrc",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TrackDetails"
                        }
                    },
                    "404": {
                        "description": "Track not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.TrackDetails": {
            "type": "object",
            "properties": {
                "artist_name": {
                    "type": "string"
                },
                "isrc": {
                    "type": "string"
                },
                "spotify_image": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}